Glasgow City Boundary and Census

Glasgow City ESRI Shapefiles imported as an `sf` object.

data("glasgowshp")

Format

A data frame with 0 observations on the following 2 variables.

x

a numeric vector

y

a numeric vector

Details

This data is provided with the support of the ESRC and JISC and uses boundary material which is copyright of the Crown, the Post Office and the ED-LINE consortium.

Source

[UK Data Service Census Support](https://census.ukdataservice.ac.uk/) with OGL

References

Examples

data(glasgowshp)
str(glasgow_pop)
#> Classes ‘sf’ and 'data.frame':	5486 obs. of  21 variables:
#>  $ oac_sub_gr: chr  "Deprived Neighbourhoods" "Constrained Commuters" "Migrant Commuters" "Endeavouring Flat Dwellers" ...
#>  $ oac_group_: chr  "Constrained Flat Dwellers" "Ethnic Dynamics" "Comfortable Cosmopolitans" "Constrained Flat Dwellers" ...
#>  $ oac_super_: chr  "Constrained City Dwellers" "Ethnicity Central" "Cosmopolitans" "Constrained City Dwellers" ...
#>  $ oac_sub__1: chr  "7b2" "3c2" "2c2" "7b3" ...
#>  $ oac_group : chr  "7b" "3c" "2c" "7b" ...
#>  $ oac_supe_1: num  7 3 2 7 7 2 2 2 7 2 ...
#>  $ name      : chr  "S00115392" "S00117060" "S00114627" "S00113610" ...
#>  $ label     : chr  "S12000046S00115392" "S12000046S00117060" "S12000046S00114627" "S12000046S00113610" ...
#>  $ shape_1_ar: num  4879 527 3735 8437 7148 ...
#>  $ shape_1_le: num  284.2 91.6 281.2 494.1 480.4 ...
#>  $ northing  : chr  "665101" "667347" "662860" "669568" ...
#>  $ easting   : chr  "253879" "261927" "258255" "257960" ...
#>  $ masterpc  : chr  "G51 4NQ" "G21 4QF" "G42 7RR" "G23 5DA" ...
#>  $ hect      : num  0.4879 0.0527 0.3735 0.8437 0.7148 ...
#>  $ sqkm      : num  0.004879 0.000527 0.003735 0.008437 0.007148 ...
#>  $ council   : chr  "S12000046" "S12000046" "S12000046" "S12000046" ...
#>  $ popcount  : num  75 82 96 85 89 81 70 93 58 60 ...
#>  $ hhcount   : num  56 43 80 47 39 43 24 52 38 29 ...
#>  $ code      : chr  "S00115392" "S00117060" "S00114627" "S00113610" ...
#>  $ objectid  : num  25749 21383 24016 21394 25459 ...
#>  $ geometry  :List of 5486
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 253941 253901 253842 253855 253864 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 261939 261938 261914 261914 261914 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 258253 258251 258230 258225 258233 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 258068 258049 258033 258030 258022 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 258081 258068 258040 257940 257870 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:17, 1:2] 258340 258310 258303 258285 258287 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 258441 258817 258827 258829 258424 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:17, 1:2] 261520 261520 261510 261501 261493 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 261083 261082 261071 261067 261064 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 261493 261490 261482 261477 261475 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 258629 258598 258597 258600 258620 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 257391 257417 257413 257409 257412 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 264620 264623 264593 264560 264543 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 256930 256920 256898 256905 256910 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 256226 256228 256216 256007 256009 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 254893 254888 254858 254818 254833 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 251980 251976 251946 251867 251843 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 252492 252479 252433 252423 252401 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 251621 251637 251664 251711 251682 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num [1:5, 1:2] 260139 260122 260050 260031 260139 ...
#>   .. .. ..$ : num [1:6, 1:2] 260081 260117 260122 260084 260083 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 261295 261294 261291 261277 261269 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 257441 257435 257402 257381 257388 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 257774 257713 257710 257737 257738 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 263032 263102 263016 262987 262961 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 264271 264212 264198 264000 263984 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 259821 259783 259763 259746 259719 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:12, 1:2] 254640 254573 254573 254568 254627 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 257259 257260 257226 257203 257234 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 255782 255743 255708 255754 255757 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 251534 251528 251340 251345 251350 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 261545 261538 261473 261397 261401 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:14, 1:2] 257678 257662 257606 257543 257523 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:5, 1:2] 253666 253650 253600 253644 253666 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num [1:8, 1:2] 254685 254687 254687 254687 254664 ...
#>   .. .. ..$ : num [1:5, 1:2] 254684 254683 254667 254668 254684 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num [1:9, 1:2] 254505 254511 254453 254447 254441 ...
#>   .. .. ..$ : num [1:6, 1:2] 254498 254476 254477 254479 254501 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 258574 258575 258545 258545 258546 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 252180 252178 252171 252150 252152 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 261140 261117 261099 261039 261023 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 260684 260682 260679 260654 260654 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 252723 252757 252800 252822 252848 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 252678 252502 252481 252547 252504 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:17, 1:2] 252951 252985 253003 253021 253024 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:14, 1:2] 255868 255868 255842 255842 255842 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 264378 264378 264442 264532 264531 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 266962 266976 267007 267016 267048 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 263645 263645 263632 263612 263613 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:12, 1:2] 255332 255322 255318 255331 255342 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:16, 1:2] 263590 263596 263592 263593 263596 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:16, 1:2] 258422 258459 258462 258459 258468 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 254276 254275 254270 254239 254245 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:16, 1:2] 258043 258043 258041 258056 258053 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 261108 261220 261258 261247 261233 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 256176 256201 256195 256174 256169 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 254757 254759 254685 254698 254745 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 257812 257806 257805 257838 257856 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:12, 1:2] 255371 255345 255278 255312 255322 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 259824 259822 259814 259770 259782 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:8, 1:2] 257818 257791 257781 257762 257734 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 258840 258819 258817 258817 258839 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 258263 258308 258304 258295 258287 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:384, 1:2] 267802 267841 267858 267880 267912 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 259808 259830 259817 259820 259792 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 255312 255278 255255 255246 255244 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:439, 1:2] 251934 252012 251936 251945 251974 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 255147 255125 255085 255091 255129 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 254274 254239 254216 254254 254302 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 254606 254534 254445 254441 254464 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 251814 251800 251769 251817 251825 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 252983 252961 252858 252881 252837 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:7, 1:2] 258279 258326 258332 258317 258239 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 260063 260069 260000 259982 259930 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 258776 258882 258848 258844 258841 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 260461 260442 260427 260305 260327 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num [1:6, 1:2] 254562 254564 254540 254538 254537 ...
#>   .. .. ..$ : num [1:6, 1:2] 254560 254559 254544 254545 254545 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 254165 254132 254109 254104 254082 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 256254 256245 256247 256246 256217 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:158, 1:2] 252879 253043 253087 253252 253405 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 259468 259466 259435 259322 259330 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:14, 1:2] 255457 255456 255455 255454 255428 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 252448 252460 252499 252525 252513 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 252282 252280 252346 252348 252304 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:10, 1:2] 267714 267724 267483 267516 267505 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 253207 253236 253183 253105 253176 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 258301 258293 258252 258229 258236 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 258448 258448 258425 258425 258425 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:9, 1:2] 263821 263821 263861 263687 263640 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 259806 259757 259752 259686 259680 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 261382 261382 261345 261354 261355 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:6, 1:2] 265373 265316 265327 265333 265372 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:11, 1:2] 267873 267860 267781 267721 267670 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:17, 1:2] 260387 260403 260463 260473 260482 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:13, 1:2] 252396 252412 252463 252467 252455 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num [1:6, 1:2] 259628 259628 259586 259586 259586 ...
#>   .. .. ..$ : num [1:5, 1:2] 259620 259596 259596 259619 259620 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:15, 1:2] 256032 256058 256076 256007 256024 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:14, 1:2] 252757 252721 252696 252678 252579 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:16, 1:2] 268247 268251 268268 268280 268419 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:14, 1:2] 253066 253045 253000 252913 252908 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:16, 1:2] 254907 254891 254839 254841 254816 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   ..$ :List of 1
#>   .. ..$ :List of 1
#>   .. .. ..$ : num [1:12, 1:2] 254715 254685 254675 254663 254667 ...
#>   .. ..- attr(*, "class")= chr [1:3] "XY" "MULTIPOLYGON" "sfg"
#>   .. [list output truncated]
#>   ..- attr(*, "n_empty")= int 0
#>   ..- attr(*, "crs")=List of 2
#>   .. ..$ input: chr "OSGB 1936 / British National Grid"
#>   .. ..$ wkt  : chr "PROJCRS[\"OSGB 1936 / British National Grid\",\n    BASEGEOGCRS[\"OSGB 1936\",\n        DATUM[\"OSGB 1936\",\n "| __truncated__
#>   .. ..- attr(*, "class")= chr "crs"
#>   ..- attr(*, "class")= chr [1:2] "sfc_MULTIPOLYGON" "sfc"
#>   ..- attr(*, "precision")= num 0
#>   ..- attr(*, "bbox")= 'bbox' Named num [1:4] 250424 656463 270434 673071
#>   .. ..- attr(*, "names")= chr [1:4] "xmin" "ymin" "xmax" "ymax"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:20] "oac_sub_gr" "oac_group_" "oac_super_" "oac_sub__1" ...
plot(glasgow_pop) # but highly recomended to use `plot(glasgow_pop[1])` and replace the numbers
#> Error in data.matrix(x): 'list' object cannot be coerced to type 'double'